`ct/l1`: move `l1_reader_test` to `l1_reader_fixture`
I'll want to use this elsewhere in a future commit.
