+++ src/v/cloud_topics/level_one/frontend_reader/tests/BUILD
@@ -1,4 +1,27 @@
-load("//bazel:test.bzl", "redpanda_cc_gtest")
+load("//bazel:test.bzl", "redpanda_cc_gtest", "redpanda_test_cc_library")
+
+redpanda_test_cc_library(
+    name = "l1_reader_fixture",
+    hdrs = ["l1_reader_fixture.h"],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//src/v/bytes:iostream",
+        "//src/v/cloud_topics:log_reader_config",
+        "//src/v/cloud_topics:logger",
+        "//src/v/cloud_topics/level_one/common:abstract_io",
+        "//src/v/cloud_topics/level_one/common:fake_io",
+        "//src/v/cloud_topics/level_one/common:object",
+        "//src/v/cloud_topics/level_one/common:object_id",
+        "//src/v/cloud_topics/level_one/frontend_reader:reader",
+        "//src/v/cloud_topics/level_one/metastore:simple",
+        "//src/v/container:chunked_circular_buffer",
+        "//src/v/container:chunked_hash_map",
+        "//src/v/model",
+        "//src/v/model/tests:random",
+        "//src/v/test_utils:gtest",
+        "@googletest//:gtest",
+    ],
+)
 
 redpanda_cc_gtest(
     name = "reader_test",
@@ -7,6 +30,7 @@ redpanda_cc_gtest(
         "reader_test.cc",
     ],
     deps = [
+        ":l1_reader_fixture",
         "//src/v/bytes:iostream",
         "//src/v/cloud_topics:log_reader_config",
         "//src/v/cloud_topics:logger",
